Fix installation on redmine 1.1.3
Installer checks for redmine version by looking at the changelog for a
particular date and parses out the version from the same line. The
broken code used to look for a particular date (probably corresponding
ot the previous release). I changed it to look at any YYYY-MM-DD string
so that it's future proof.