Corrected previous taskboard color fix for groups.
Turned out task.js was not supposed to be altered. Colors of tasks being moved
turned grey because the previous fix assumed always to find the selected color
in the assigned_to_id-select.
The select option has now a default color added for groups and the task.js is
altered to use find() in stead of children() in the beforeSave function. This
allows finding selected group nested inside a "<optgroup>".