Add a setting to always show task priority.
The task priority current shows up on the taskboard view if it is
anything other than 'Normal'. It also shows up in the task editing popup
under those conditions.
We want to be able to change the task priority from the taskboard, which
is not currently possible if the starting priority is 'Normal'.

It seemed that the easiest way to solve this was to add an option to
always show the task priority. This has worked well for us for several
months now.