Initial CRUD support for creating "Release Multiview" inspired by http://blog.gdinwiddie.com/2012/12/04/multi-release-burnup/.
The idea is to put multiple releases into a burnup chart to get an
estimate of when a series of releases is finished according to the
velocity of all releases combined.

The user has to choose which releases to put into the multiview.
It might be interesting to see a series of sequential releases done
by the same team. Another possibility is to select parallel releases
from different teams if that makes sense in the given project.

As such the "Release Multiview" will just allow to group a selection
of releases visible in a given project's scope and generate a graph
with the combined release burnups.